Add files via upload
Please share it with everyone, please copy this page and would be extremely helpful if you would set it on any of your servers. This is about awareness of society. Subject is too technical and too difficult to understand for most of people. So slowly making them aware of what might be going on, may be quite good idea. These people rely on opinion of others or popularity. If topic will gain popularity, at least to some point where some not only solutions but also education will happen. I consider this as ultra necessary unless we like to wake up always late and cry that well... indeed this might be a HUGE problem. Quotation I chose were the most clear as they could ever been. Cheers. :)